kb/requirements.txt
streamlit
langchain
openai
chromadb
python-dotenv
requests
tqdm
pypdf
